aligned jearly/monthly label with right edge

+++ app/src/main/java/ch/band/inf2019/uk335/activities/OverviewActivity.java
@@ -37,7 +37,7 @@ public abstract class OverviewActivity  extends AppCompatActivity {
         } else {
             textViewFrequency.setText(getString(R.string.monthly));
         }
-        int cost =viewModel.getCostMonthYear()/100;
+        double cost =viewModel.getCostMonthYear()/100;
         textViewSum.setText(NumberFormat.getCurrencyInstance().format(cost));
     }
 }
